LayoutTests/platform/wpe/svg/foreignObject/fO-fixed-position-crash-expected.txt

CONSOLE MESSAGE: InvalidStateError: execCommand is only supported on HTML documents.
